bus/dpaa: Add shared mac support
A shared MAC interface is an interface which can be used
by both kernel and userspace based on classification configuration
It is defined in dts with the compatible string "fsl,dpa-ethernet-shared"
which bpool will be seeded by the dpdk partition and configured
as a netdev by the dpaa Linux eth driver.
User space buffers from the bpool will be kmapped by the kernel.

Signed-off-by: Radu Bulie <radu-andrei.bulie@nxp.com>
DPDK-1798

Conflicts:
	drivers/mempool/dpaa/Makefile