PVT: mk: remove pedantic from compile error list
While compiling on Flexbuild, because of some compiler/glibc
versions, following error is printed for DPDK compilation:

    epoll.h:64:15: error: ISO C restricts enumerator values to \
    range of ‘int’ [-Werror=pedantic]

This is reported for usr/include/aarch64-linux-gnu/sys/epoll.h
file while compiling failsafe driver.

This error lies outside the scope of DPDK (as this is in a glibc
provided header file), but a hotfix has been enabled in DPDK to
enable Flexbuild.

This should be removed in future.

Signed-off-by: Shreyansh Jain <shreyansh.jain@nxp.com>